refactor commands to use central delete function
