+++ src/commands/fileShare/directoryActionHandlers.ts
@@ -6,9 +6,10 @@
 import { AzExtParentTreeItem, IActionContext, registerCommand } from '@microsoft/vscode-azext-utils';
 import { DirectoryTreeItem } from '../../tree/fileShare/DirectoryTreeItem';
 import { IFileShareCreateChildContext } from '../../tree/fileShare/FileShareTreeItem';
+import { deleteFilesAndDirectories } from '../deleteFilesAndDirectories';
 
 export function registerDirectoryActionHandlers(): void {
-    registerCommand("azureStorage.deleteDirectory", async (context: IActionContext, treeItem: AzExtParentTreeItem) => await treeItem.deleteTreeItem(context));
+    registerCommand("azureStorage.deleteDirectory", deleteFilesAndDirectories);
     registerCommand("azureStorage.createSubdirectory", async (context: IActionContext, treeItem: AzExtParentTreeItem) => await treeItem.createChild(<IFileShareCreateChildContext>{ ...context, childType: DirectoryTreeItem.contextValue }));
 
     // Note: azureStorage.createTextFile is registered in fileShareActionHandlers
